requirements.txt
pynvim
openai
tiktoken
requests
markdownify
bs4
